Refactor: Improve site structure, UI, and viewers
This commit addresses multiple issues aimed at improving the project's
maintainability, user interface, and code readability.

Key changes include:

1.  **Project Structure Refactoring**:
    *   Generator script `generate-pages.js` moved to `src/generator/main.js`
        and broken down into multiple modules (`config.js`, `html-generator.js`,
        `file-processor.js`, `static-handler.js`, `file-utils.js`) for
        better organization.
    *   CSS, JavaScript, and HTML templates moved into `src/assets/css`,
        `src/assets/js`, and `src/templates` respectively.
    *   `package.json` updated to reflect new script paths.

2.  **UI and Styling Enhancements**:
    *   Button styling improved across the site for a more modern and
        consistent look and feel, addressing "botones que no se ven bien".
    *   General UI elements like directory listings (`tech-box`) and tables
        (`comparison-table`) enhanced with better spacing and hover effects.
    *   Viewer styles (PDF, Word, PowerPoint) standardized using a common
        CSS file (`viewer-common.css`) and aligned with the main site theme.

3.  **Viewer Improvements**:
    *   Inline CSS and JavaScript extracted from Word and PowerPoint viewer
        templates into external files.
    *   PDF viewer controls reorganized for better usability and responsiveness.
        A download button was added to the PDF viewer.
    *   Paths in HTML templates and configuration updated to reflect new
        asset locations.

4.  **Documentation**:
    *   A comprehensive `README.md` created, detailing the new project
        structure, build/development process, and purpose of key scripts.

The build process has been verified after these changes. A manual testing
checklist has been prepared to ensure all functionalities and UI aspects
are working as expected.
