+++ README.md
@@ -1,0 +1,83 @@
+# Institute Documentation Site
+
+This project is a static website that serves educational materials and resources for the institute.
+
+## Overview
+
+The website is generated by a Node.js script that processes files from the `documents/` directory and creates a structured set of HTML pages in the `docs/` directory.
+
+Key features include:
+- Automatic generation of navigation pages for directories.
+- Viewers for PDF, Word (.docx), and PowerPoint (.pptx) file types.
+- Course timetables.
+
+## Project Structure
+
+*   `documents/`: Contains the source material (PDFs, DOCX, images, etc.) organized by subject and year. This is the content source for the website.
+*   `src/`: Contains all the source code for the static site generator and its assets.
+    *   `src/generator/`: Node.js scripts responsible for generating the site.
+        *   `main.js`: The main entry point for the build process. Orchestrates the site generation.
+        *   `config.js`: Holds the main configuration object (`CONFIG`) for the generator, including paths and file processing options.
+        *   `html-generator.js`: Responsible for creating the HTML structure for directory listing pages.
+        *   `file-processor.js`: Handles the logic for traversing source directories, processing files, and managing content for HTML pages.
+        *   `static-handler.js`: Manages the copying of static assets (CSS, JS, images, templates) to the output directory and generation of specific pages like viewers.
+        *   `file-utils.js`: Contains utility functions for file name formatting, exclusion checks, and determining file types for viewers.
+    *   `src/assets/css/`: Contains all source CSS files.
+        *   `styles.css`: Main stylesheet for the website's overall look and feel.
+        *   `pdf-viewer.css`: Specific styles for the PDF viewer page.
+        *   `viewer-common.css`: Common styles for other document viewers (Word, PowerPoint).
+    *   `src/assets/js/`: Contains all source JavaScript files used by the website.
+        *   `modal.js`: Handles modal dialog functionality.
+        *   `pdf-viewer.js`: Core logic for the PDF viewer.
+        *   `word-viewer.js`: Logic for the Word document viewer.
+        *   `powerpoint-viewer.js`: Logic for the PowerPoint presentation viewer.
+    *   `src/templates/`: HTML template files used for generating specific pages like the document viewers.
+        *   `pdf-viewer.html`: Template for the PDF viewer.
+        *   `word-viewer.html`: Template for the Word viewer.
+        *   `powerpoint-viewer.html`: Template for the PowerPoint viewer.
+*   `docs/`: The output directory for the generated static website. This directory is what gets deployed and served to users.
+*   `package.json`: Node.js project metadata, dependencies, and script definitions.
+*   `README.md`: This file, providing information about the project.
+
+## Development
+
+This project uses Node.js and pnpm (or npm) for dependency management and running scripts.
+
+**Prerequisites:**
+*   Node.js (LTS version recommended, e.g., 18.x or 20.x)
+*   pnpm (recommended, or npm)
+
+**Installation:**
+To install dependencies:
+```bash
+pnpm install
+```
+or if using npm:
+```bash
+npm install
+```
+
+**Building the Site:**
+To generate the static site into the `docs/` directory:
+```bash
+pnpm build
+```
+or
+```bash
+npm run build
+```
+This command executes the `src/generator/main.js` script.
+
+**Running the Development Server:**
+To build the site and serve it locally (usually on `http://localhost:3000` as per `serve` default):
+```bash
+pnpm dev
+```
+or
+```bash
+npm run dev
+```
+This is useful for previewing changes as you develop. The `serve` command (from `devDependencies`) will host the `docs/` directory.
+The specific port for the dev server might vary if port 3000 is in use. Check the output of the `pnpm dev` or `npm run dev` command.
+
+This will typically build the site and serve the `docs/` directory.
