--- css/pdf-viewer.css
@@ -1,131 +1,0 @@
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  box-sizing: border-box;
-}
-
-#pdf-container {
-  width: 100%;
-  max-width: 800px;
-  margin: 20px auto;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-#pdf-viewer {
-  border: 1px solid #ccc;
-  margin-bottom: 10px;
-  flex: 1;
-  overflow: auto;
-  position: relative;
-  background: #f5f5f5;
-  cursor: grab;
-}
-
-#pdf-viewer:active {
-  cursor: grabbing;
-}
-
-.controls {
-  display: flex;
-  gap: 10px;
-  margin: 10px 0;
-  align-items: center;
-  flex-wrap: wrap;
-  justify-content: center;
-  background: white;
-  padding: 10px;
-  border-radius: 4px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-#page-num {
-  width: 50px;
-  text-align: center;
-  padding: 4px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-canvas {
-  max-width: none;
-  height: auto;
-  display: block;
-  margin: 0 auto;
-}
-
-button {
-  padding: 8px 16px;
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-button:hover {
-  background-color: #45a049;
-}
-
-.zoom-controls,
-.fit-controls {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.zoom-controls button {
-  padding: 4px 8px;
-  font-size: 16px;
-  font-weight: bold;
-}
-
-#zoom-level {
-  min-width: 60px;
-  text-align: center;
-}
-
-.fit-controls button {
-  padding: 4px 8px;
-  font-size: 16px;
-}
-
-#fullscreen {
-  margin-left: auto;
-  background-color: #2196f3;
-}
-
-#fullscreen:hover {
-  background-color: #0b7dda;
-}
-
-#pdf-container.fullscreen {
-  max-width: 100%;
-  margin: 0;
-  padding: 20px;
-  height: 100vh;
-  width: 100vw;
-  box-sizing: border-box;
-}
-
-#pdf-container.fullscreen #pdf-viewer {
-  max-height: calc(100vh - 100px);
-}
-
-/* Estilos para el modo de ajuste */
-#pdf-viewer.fit-width canvas {
-  width: 100% !important;
-  height: auto !important;
-}
-
-#pdf-viewer.fit-height canvas {
-  height: 100% !important;
-  width: auto !important;
-}
