+++ docs/word-viewer.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Visor de Documentos Word | Instituto 166</title>
+  <link rel="stylesheet" href="css/styles.css">
+  <link rel="stylesheet" href="css/viewer-common.css">
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/mammoth/1.6.0/mammoth.browser.min.js"></script>
+</head>
+<body>
+  <div class="container">
+    <div class="header">
+      <h1>Visor de Documentos Word</h1>
+      <div>
+        <button id="back-btn" onclick="window.history.back()">⬅ Volver</button>
+        <button onclick="window.location.href='index.html'">🏠 Inicio</button>
+      </div>
+    </div>
+    
+    <div class="file-info" id="file-info"></div>
+    
+    <div class="controls">
+      <button id="download-btn">📥 Descargar Original</button>
+    </div>
+    
+    <div class="doc-viewer" id="doc-content">
+      <div class="loading">Cargando documento...</div>
+    </div>
+  </div>
+
+  <script src="scripts/word-viewer.js"></script>
+</body>
+</html>
