--- pdf-viewer.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0"
-    />
-    <title>Visor de PDF</title>
-    <link rel="stylesheet" href="../css/pdf-viewer.css" />
-  </head>
-  <body>
-    <div id="pdf-container">
-      <div id="pdf-viewer"></div>
-      <div class="controls">
-        <button id="prev">Anterior</button>
-        <span>
-          Página
-          <input
-            type="number"
-            id="page-num"
-            min="1"
-            value="1"
-          />
-          de
-          <span id="page-count">0</span>
-        </span>
-        <button id="next">Siguiente</button>
-        <div class="zoom-controls">
-          <button id="zoom-out" title="Reducir">-</button>
-          <span id="zoom-level">100%</span>
-          <button id="zoom-in" title="Ampliar">+</button>
-        </div>
-        <div class="fit-controls">
-          <button id="fit-width" title="Ajustar al ancho">↔</button>
-          <button id="fit-height" title="Ajustar al alto">↕</button>
-        </div>
-        <button id="fullscreen">Pantalla completa</button>
-      </div>
-    </div>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.11.174/pdf.min.js"></script>
-    <script src="../scripts/pdf-viewer.js"></script>
-  </body>
-</html>
