+++ src/assets/css/styles.css
@@ -40,6 +40,29 @@ html {
   /* Quitar height: 100% si no se necesita footer pegado al fondo siempre */
 }
 
+button, .action-button, .calendar-button { /* Add other common button classes if any */
+  display: inline-block;
+  border: 1px solid transparent;
+  padding: 0.6em 1.2em; /* Adjust for comfortable size */
+  margin: 0.3em; /* Add some margin for spacing */
+  font-size: 0.95rem; /* Ensure readable text */
+  font-weight: 500;
+  line-height: 1.4;
+  text-align: center;
+  text-decoration: none;
+  white-space: nowrap;
+  vertical-align: middle;
+  cursor: pointer;
+  user-select: none;
+  border-radius: 5px; /* Softer edges */
+  transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+  /* Remove default browser styling if needed */
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+}
+
+
 body {
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   font-size: 16px;
@@ -184,6 +207,13 @@ h3 {
   display: flex;
   flex-direction: column;
   min-height: auto; /* Altura automática según contenido */
+}
+.tech-box:hover {
+  box-shadow: var(--shadow-hover); /* Use existing variable if suitable, or define a new one */
+  transform: translateY(-2px); /* Optional: slight lift effect */
+}
+.tech-box-link { /* Ensure the link wrapper has no decoration if the box provides all visuals */
+    text-decoration: none;
 }
 .tech-box > *:last-child {
     margin-bottom: 0; /* Quitar margen inferior del último elemento */
@@ -459,15 +489,74 @@ footer .footer-bar {
 footer .footer-bar button {
   background: rgba(255, 255, 255, 0.1);
   color: white;
-  padding: 0.6rem 1rem;
-  border-radius: 4px;
+  /* padding: 0.6rem 1rem; // Harmonized by general button style */
+  /* border-radius: 5px; // Harmonized by general button style */
   border: 1px solid rgba(255, 255, 255, 0.2);
-  cursor: pointer;
-  transition: var(--transition);
+  /* transition: var(--transition); // Harmonized by general button style */
 }
-footer .footer-bar button:hover {
+footer .footer-bar button:hover, footer .footer-bar button:focus {
   background: rgba(255, 255, 255, 0.2);
+  box-shadow: 0 1px 3px rgba(0,0,0,0.1); /* Optional subtle shadow */
 }
+
+/* Specific styles for .action-button */
+.action-button {
+  background-color: var(--primary);
+  color: white;
+  border-color: var(--primary);
+}
+.action-button:hover, .action-button:focus {
+  background-color: var(--primary-dark);
+  border-color: var(--primary-dark);
+  box-shadow: 0 2px 5px rgba(0,0,0,0.1);
+}
+
+/* Specific styles for .calendar-button */
+.calendar-button {
+  background-color: var(--secondary); /* Using secondary for distinction */
+  color: white; /* Assuming secondary is dark enough, otherwise adjust */
+  border-color: var(--secondary);
+}
+.calendar-button:hover, .calendar-button:focus {
+  background-color: #2aaadd; /* Slightly darker shade of secondary */
+  border-color: #2aaadd;
+  box-shadow: 0 2px 5px rgba(0,0,0,0.1);
+}
+
+/* Comparison Table Enhancements */
+.comparison-table {
+  width: 100%; /* Ensure table takes full width */
+  border-collapse: collapse; /* Cleaner borders */
+  margin-bottom: 1.5rem; /* Space below table */
+}
+.comparison-table th, .comparison-table td {
+  padding: 0.8rem 1rem; /* Increase padding for better spacing */
+  text-align: left;
+  border-bottom: 1px solid var(--border);
+  border-left: 1px solid var(--border-light); /* Add subtle vertical lines */
+  border-right: 1px solid var(--border-light);
+}
+.comparison-table th:first-child, .comparison-table td:first-child {
+    border-left: none; /* Remove leftmost border for cleaner look */
+}
+.comparison-table th:last-child, .comparison-table td:last-child {
+    border-right: none; /* Remove rightmost border for cleaner look */
+}
+.comparison-table th {
+  background-color: var(--border-light); /* Light background for headers */
+  font-weight: 600;
+  color: var(--primary-dark);
+  border-top: 1px solid var(--border); /* Add top border to header */
+}
+.comparison-table tr:last-child td {
+    border-bottom: 1px solid var(--border); /* Ensure last row has a bottom border */
+}
+/* Optional: hover effect for rows - can be distracting
+.comparison-table tbody tr:hover {
+  background-color: #f0f8ff; 
+}
+*/
+
 
 /* --- AJUSTES RESPONSIVOS (Revisados) --- */
 @media (max-width: 992px) {
