+++ src/assets/css/viewer-common.css
@@ -1,0 +1,95 @@
+/* CSS rules for Word and PowerPoint viewers */
+body {
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif; /* Match main styles.css */
+  margin: 0;
+  padding: 0;
+  background-color: #f8f9fa; /* var(--light) from styles.css */
+  color: #333333; /* var(--text) from styles.css */
+}
+.container {
+  max-width: 900px;
+  margin: 20px auto; /* Added top/bottom margin */
+  padding: 20px;
+  background-color: white; /* Explicit white background for content area */
+  border-radius: 8px; /* Softer radius */
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08); /* var(--shadow) */
+}
+.doc-viewer, .presentation-viewer {
+  background-color: white;
+  border: 1px solid #dee2e6; /* var(--border) */
+  border-radius: 5px;
+  padding: 30px; /* Slightly reduced padding */
+  min-height: 350px; /* Slightly reduced min-height */
+  box-shadow: none; /* Remove redundant shadow if container has it */
+}
+.header { /* This is the header within the viewer page, not the main site header */
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
+  padding-bottom: 15px; /* Added padding */
+  border-bottom: 1px solid #e9ecef; /* var(--border-light) */
+}
+.header h1 {
+  margin: 0;
+  font-size: 1.6rem; /* Consistent with main H1 if possible, or slightly smaller */
+  color: #2980b9; /* var(--primary-dark) */
+  font-weight: 600;
+}
+.loading {
+  text-align: center;
+  padding: 50px 0;
+  font-style: italic;
+  color: #6c757d; /* var(--text-light) */
+}
+.error {
+  color: #e74c3c; /* var(--accent) for error text */
+  border: 1px solid #e74c3c;
+  padding: 15px; /* More padding */
+  border-radius: 5px;
+  background-color: #fdedec; /* Lighter red background */
+}
+.controls {
+  margin-bottom: 20px;
+  display: flex; /* Added for button alignment */
+  gap: 10px; /* Space between buttons */
+}
+
+/* Unified Button Styling - inspired by styles.css .action-button and general button */
+button, .button { /* Using a class .button for more flexibility if needed */
+  padding: 0.6rem 1rem; /* Match footer button style */
+  background-color: #3498db; /* var(--primary) */
+  color: white;
+  border: 1px solid transparent; /* For consistency, can be var(--primary-dark) */
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 0.95rem; /* Slightly larger */
+  text-decoration: none; /* For <a> styled as buttons */
+  display: inline-flex; /* To align icons if any */
+  align-items: center;
+  gap: 0.4em;
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1); /* var(--transition) */
+  font-weight: 500; /* Medium weight */
+}
+button:hover, .button:hover {
+  background-color: #2980b9; /* var(--primary-dark) */
+  box-shadow: 0 2px 8px rgba(0,0,0,0.1); /* Subtle shadow on hover */
+}
+/* Specific button for "Back" or "Home" to be less prominent */
+button.secondary, .button.secondary {
+  background-color: #6c757d; /* var(--text-light) or a gray */
+  border-color: #6c757d;
+}
+button.secondary:hover, .button.secondary:hover {
+  background-color: #5a6268; /* Darker gray */
+}
+
+.file-info {
+  margin-bottom: 15px; /* More space */
+  font-size: 0.9rem;
+  color: #6c757d; /* var(--text-light) */
+  padding: 10px;
+  background-color: #f8f9fa; /* Light background for info */
+  border-radius: 4px;
+  border: 1px solid #e9ecef;
+}
